gradio/langchain-vertex-test/requirements.txt
gradio
langchain-google-vertexai
langchain_core
python-dotenv